Show entire message bodies.
Fix a bug where the height for the table row for message bodies was not sized appropriately for the content.
Also try harder to get the message body by digging into the message parts recursively. There are still messages which generate blank (apparently very long) message bodies. Not sure why but it's beyond the scope of this example to parse message data beyond base64url decoding.